% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{buildContributionP2Path}
\alias{buildContributionP2Path}
\title{Build Contribution Pillar II Path}
\usage{
buildContributionP2Path(birthday, Salary, SalaryGrowthRate, CurrentP2,
  P2purchase, TypePurchase, rate = BVGMindestzinssatz,
  givenday = today("UTC"), RetirementAge)
}
\arguments{
\item{birthday}{date of birth ("YYYY-MM-DD").}

\item{Salary}{stipend}

\item{SalaryGrowthRate}{rate at which the salary increases/decreases}

\item{CurrentP2}{value of the current assets in the Occupational Pension Fund.}

\item{P2purchase}{annual voluntary contribution to the Occupational Pension Fund.}

\item{TypePurchase}{AnnualP2 if the purchase on the pillar II gets repeated every year until retirement.}

\item{rate}{male or female}

\item{givenday}{date at which the age is computed ("YYYY-MM-DD").}

\item{RetirementAge}{age of retirement}

\item{ncp}{length contribution path to retirement}
}
\value{
data frame with annual different contributions to the Pillar II.
}
\description{
Gather all the required information to project the annual contributions to the occupational pension fund.
}
\examples{
\dontrun{
buildContributionP2Path(
  birthday = "1975-10-10",
  Salary = 90000,
  SalaryGrowthRate = 0.01,
  CurrentP2 = 10000,
  P2purchase = 2000,
  TypePurchase = "AnnualP2",
  rate = 0.025,
  givenday = as.Date("2018-07-04"),
  RetirementAge = 67)
}
}
