% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{buildContributionP3path}
\alias{buildContributionP3path}
\title{Build Contribution Pillar III Path}
\usage{
buildContributionP3path(birthday, P3purchase, CurrentP3, returnP3,
  givenday = today("UTC"), RetirementAge)
}
\arguments{
\item{birthday}{date of birth ("YYYY-MM-DD").}

\item{P3purchase}{annual voluntary contribution to the Private Pension Fund (Pillar 3).}

\item{CurrentP3}{value of the current assets in the Private Pension Fund (Pillar 3).}

\item{returnP3}{annual expected return on the Private Pension Fund.}

\item{givenday}{date at which the age is computed ("YYYY-MM-DD").}

\item{RetirementAge}{age of retirement}
}
\value{
data frame with annual different contributions to the Pillar III.
}
\description{
Build the contribution path for a standard pension fund, called Pillar III in Switzerland.
}
\examples{
\dontrun{
buildContributionP3path(
  birthday = "1980-12-01",
  P3purchase = 5000,
  CurrentP3 = 100000,
  returnP3 = 0.03,
  givenday = as.Date("2015-11-30"),
  RetirementAge = 62)
}
}
