% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{buildt}
\alias{buildt}
\title{Buildt}
\usage{
buildt(birthday, givenday = today("UTC"), RetirementAge = 65)
}
\arguments{
\item{birthday}{date of birth ("YYYY-MM-DD").}

\item{givenday}{date at which the age is computed ("YYYY-MM-DD").}

\item{RetirementAge}{age of retirement}
}
\value{
TODO-Gabriel
}
\description{
TODO-Gabriel
}
\examples{
\dontrun{
buildt("1981-08-12")
}
}
