% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxBenefit.R
\name{buildTaxBenefits}
\alias{buildTaxBenefits}
\title{Builds a data frame with the tax benefits path}
\usage{
buildTaxBenefits(birthday, TypePurchase, P2purchase, P3purchase, returnP3,
  Salary, SalaryGrowthRate, postalcode, NKids, churchtax, rate_group,
  givenday = today("UTC"), RetirementAge)
}
\arguments{
\item{birthday}{date of birth ("YYYY-MM-DD").}

\item{TypePurchase}{AnnualP2 if the purchase on the pillar II gets repeated every year until retirement.}

\item{P2purchase}{annual voluntary contribution to the Occupational Pension Fund.}

\item{P3purchase}{annual voluntary contribution to the Private Pension Fund (Pillar 3).}

\item{returnP3}{annual expected return on the Private Pension Fund.}

\item{Salary}{stipend}

\item{SalaryGrowthRate}{rate at which the salary increases/decreases}

\item{postalcode}{zip code \code{Character}}

\item{NKids}{number of children \code{Numeric} scalar.}

\item{churchtax}{Y/N \code{Character} Y/N}

\item{rate_group}{A (Single), B (Married), C (Married Double income) \code{Character}.}

\item{givenday}{date at which the age is computed ("YYYY-MM-DD").}

\item{RetirementAge}{age of retirement}

\item{CurrentP2}{value of the current assets in the Occupational Pension Fund.}

\item{CurrentP3}{value of the current assets in the Private Pension Fund (Pillar 3).}

\item{ncp}{length contribution path to retirement}
}
\value{
data.frame tax benefit path.
}
\description{
Builds a data frame with the tax benefits path
}
\details{
All inputs are scalars. Builds a data frame as long as the years to retirement.
Calls 'getTaxAmount()' through 'calcTaxBenefitSwiss()', therefore, it assumes objects on the global enviornment.
}
\examples{
\dontrun{buildTaxBenefits(
 birthday,
 TypePurchase,
 P2purchase,
 P3purchase,
 returnP3,
 Salary,
 SalaryGrowthRate,
 postalcode,
 NKids,
 churchtax,
 rate_group,
 MaxContrTax,
 givenday = today("UTC"),
 RetirementAge = 65)
}
}
\seealso{
Other swisstax: \code{\link{calcTaxBenefitSwiss}},
  \code{\link{getTaxAmount}}, \code{\link{lookupTaxAmount}}
}
\concept{swisstax}
