% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calcAge}
\alias{calcAge}
\title{Calculate Age}
\usage{
calcAge(birthday, givenday = today("UTC"))
}
\arguments{
\item{birthday}{date of birth ("YYYY-MM-DD").}

\item{givenday}{date at which the age is computed ("YYYY-MM-DD").}
}
\value{
calculated age in years
}
\description{
Function to calculate person's age at a specific point in time.
}
\details{
Calculated as the difference between current date and birthday.
}
\examples{
 calcAge("1981-08-12")
}
