% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calcBVGpurchase}
\alias{calcBVGpurchase}
\title{Calculate BVG Purchase}
\usage{
calcBVGpurchase(TypePurchase, P2purchase, ncp)
}
\arguments{
\item{TypePurchase}{type of purchase}

\item{P2purchase}{TODO-Gabriel}

\item{ncp}{TODO-Gabriel}
}
\value{
BVG purchase
}
\description{
TODO-Gabriel
}
\examples{
\dontrun{
calcBVGpurchase(TypePurchase = "AnnualP2", P2purchase = 2000, ncp = 25) \%>\% print
}
}
