% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calcExpectedSalaryPath}
\alias{calcExpectedSalaryPath}
\title{calcExpectedSalaryPath}
\usage{
calcExpectedSalaryPath(Salary, SalaryGrowthRate, ncp)
}
\arguments{
\item{Salary}{Stipend}

\item{SalaryGrowthRate}{Rate at which the salary increases/decreases}

\item{ncp}{Length contribution path to retirement}
}
\value{
Expected salary path.
}
\description{
Calculate whether the salary will increase/decrease and by how much.
}
\examples{
\dontrun{
calcExpectedSalaryPath(90000, 0.02, 20)
}
}
