% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxBenefit.R
\name{calcTaxBenefitSwiss}
\alias{calcTaxBenefitSwiss}
\title{Calculate Tax Benefit Swiss}
\usage{
calcTaxBenefitSwiss(ExpectedSalaryPath, TaxableIncome, rate_group, Age,
  NKids, postalcode, churchtax)
}
\arguments{
\item{ExpectedSalaryPath}{vector length equals year to retirement}

\item{TaxableIncome}{vector length equals year to retirement}

\item{rate_group}{TODO-Gabriel. \code{Character}.}

\item{Age}{Age of the person. \code{Numeric}}

\item{NKids}{number of children}

\item{postalcode}{zip code}

\item{churchtax}{TODO-Gabriel}
}
\value{
TODO-Gabriel
}
\description{
Calculates the tax benefit as a difference of the taxes paid with and without retirement contributions.
Calls 'getTaxAmount()', therefore, it assumes objects in the global environment.
}
\examples{
\dontrun{
  calcTaxBenefitSwiss(ExpectedSalaryPath = seq(90000, 100000, 1000),
                    TaxableIncome = seq(88000, 98000, 1000),
                    rate_group = "A",
                    Age = seq(55, 65),
                    NKids = 0,
                    postalcode = 8400,
                    churchtax = "Y")
}
}
\seealso{
\code{\link[=getTaxAmount]{getTaxAmount()}}

Other swisstax: \code{\link{buildTaxBenefits}},
  \code{\link{getTaxAmount}}, \code{\link{lookupTaxAmount}}
}
\concept{swisstax}
