% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{changeToPercentage}
\alias{changeToPercentage}
\title{Change to Percentage}
\usage{
changeToPercentage(df)
}
\arguments{
\item{df}{TODO-Gabriel}
}
\value{
percentage value
}
\description{
from decimal to percentage value
}
