% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_SmaRP_version.R
\name{get_SmaRP_version}
\alias{get_SmaRP_version}
\title{SmaRP version}
\usage{
get_SmaRP_version()
}
\value{
The version of SmaRP as character string.
}
\description{
Returns the version of the installed SmaRP package and app.
}
