% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{getRetirementCalendar}
\alias{getRetirementCalendar}
\title{Obtain Retirement Calendar}
\usage{
getRetirementCalendar(birthday, givenday = today("UTC"),
  RetirementAge = 65)
}
\arguments{
\item{birthday}{date of birth ("YYYY-MM-DD").}

\item{givenday}{date at which the age is computed ("YYYY-MM-DD").}

\item{RetirementAge}{age of retirement}
}
\value{
retirement calendar
}
\description{
TODO-Gabriel
}
\examples{
\dontrun{
getRetirementCalendar("1981-08-12")
}
}
