% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{getRetirementday}
\alias{getRetirementday}
\title{Get Retirement Day}
\usage{
getRetirementday(birthday, RetirementAge = 65)
}
\arguments{
\item{birthday}{date of birth ("YYYY-MM-DD").}

\item{RetirementAge}{age of retirement}
}
\value{
day at which retirement begins
}
\description{
Calculate day of retirement assuming the person can retire at the age of 65
}
\examples{
 getRetirementday("1981-08-12")
}
