% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxBenefit.R
\name{getTaxAmount}
\alias{getTaxAmount}
\title{Calculates the tax amount (federal, kantonal and local)}
\usage{
getTaxAmount(Income, rate_group, Age, NKids, postalcode, churchtax)
}
\arguments{
\item{Income}{scalar numeric}

\item{rate_group}{character}

\item{Age}{scalar numeric}

\item{NKids}{scalar numeric}

\item{postalcode}{character}

\item{churchtax}{character Y/N}
}
\description{
This function uses 2 main sources for tax data.
At Kanton and Gemeinde level, the source is taxburden.list.
At federal level, we use the official taxrate table (BundessteueTabelle) and we try to aproximate the taxable income.
}
\details{
This function assumes the following objects on the global enviornment
 * PLZGemeinden (includes Kirchensteuer)
 * taxburden.list
 * BundessteueTabelle
 * BVGcontriburionratesPath, BVGcontriburionrates
 * MaxBVG, MinBVG
 * KinderabzugKG
 * NBU, maxNBU
 * AHL
 * ALV, maxALV
 * VersicherungsL, VersicherungsV, VersicherungsK
 * BerufsauslagenTarif, BerufsauslagenMax, BerufsauslagenMin
}
\examples{
\dontrun{
getTaxAmount(Income = 200000, rate_group = "C", Age = 32, NKids = 5, postalcode = 8400, churchtax = "Y")
}
}
\seealso{
Other swisstax: \code{\link{buildTaxBenefits}},
  \code{\link{calcTaxBenefitSwiss}},
  \code{\link{lookupTaxAmount}}
}
\concept{swisstax}
