% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IB}
\alias{IB}
\title{IB list}
\format{Named \code{list} of Info Box tooltip text.}
\usage{
IB
}
\description{
Info Box tooltip text for the various inputs of the SmaRP web app.
}
\keyword{datasets}
