% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{isnotAvailable}
\alias{isnotAvailable}
\title{Is not Available}
\usage{
isnotAvailable(inputValue)
}
\arguments{
\item{inputValue}{input value}
}
\value{
TRUE or FALSE
}
\description{
if input value is not available, then return logical TRUE else FALSE
}
