% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{isnotAvailableReturnZero}
\alias{isnotAvailableReturnZero}
\title{isnotAvailableReturnZero}
\usage{
isnotAvailableReturnZero(inputValue, fallback = 0)
}
\arguments{
\item{inputValue}{Input value.}

\item{fallback}{Zero.}
}
\value{
Zero or input value.
}
\description{
If input value is not available, then return zero else input value.
}
