% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_application.R
\name{launch_application}
\alias{launch_application}
\title{launch_application}
\usage{
launch_application(launch.browser = interactive(), ...)
}
\arguments{
\item{launch.browser}{If true, the system's default web browser will be
launched automatically after the app is started. Defaults to true in
interactive sessions only. This value of this parameter can also be a
function to call with the application's URL.}

\item{...}{Additional arguments passed on to \code{\link[=runApp]{runApp()}}.}
}
\value{
Side-effecting function. Launches the SmaRP app.
}
\description{
Launch the SmaRP Shiny wep app.
}
\examples{
if (interactive()) {
  SmaRP::launch_application()
}
}
