% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxBenefit.R
\name{lookupTaxAmount}
\alias{lookupTaxAmount}
\title{lookupTaxAmount}
\usage{
lookupTaxAmount(Income, Tabelle, CivilStatus)
}
\arguments{
\item{Income}{Annual stipend.}

\item{Tabelle}{Income - Tax rate table at Federal level.}

\item{CivilStatus}{Marital status.}
}
\value{
Tax amount to be paid.
}
\description{
Search the tax amount to be paig given one income on the tax tables.
}
\examples{
\dontrun{
lookupTaxAmount(Income = 100000, Tabelle = BundessteueTabelle, CivilStatus = "A")
}
}
\seealso{
swisstax
}
