% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{makeTable}
\alias{makeTable}
\title{makeTable}
\usage{
makeTable(Road2Retirement, moncols = c("DirectP2", "ReturnP2", "TotalP2",
  "DirectP3", "ReturnP3", "TotalP3", "DirectTax", "ReturnTax", "TotalTax",
  "Total"))
}
\arguments{
\item{Road2Retirement}{Main data frame where main results are displayed.}

\item{moncols}{Columns to prit out on the table.}
}
\value{
Table to print out.
}
\description{
Utility function to display main results on the table tab.
}
\examples{
\dontrun{
makeTable(Road2Retirement)
}

}
