% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{need_not_zero}
\alias{need_not_zero}
\title{need_not_zero}
\usage{
need_not_zero(input, inputname)
}
\arguments{
\item{input}{zero, nothing or null.}

\item{inputname}{Name of input.}
}
\value{
Warning message or nothing.
}
\description{
Utility function to display a message in case a non zero value is needed.
}
