% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{printCurrency}
\alias{printCurrency}
\title{Print Currency}
\usage{
printCurrency(value, digits = 0, sep = ",", decimal = ".")
}
\arguments{
\item{value}{currency worth}

\item{digits}{currency digit number}

\item{sep}{comma for thousands separation}

\item{decimal}{dot for decimal separation}
}
\value{
currency
}
\description{
TODO-Gabriel
}
\examples{
\dontrun{
printCurrency(123123.334)
}
}
