% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{returnPLZKanton}
\alias{returnPLZKanton}
\title{Return Postal Code Kanton}
\usage{
returnPLZKanton(plz)
}
\arguments{
\item{plz}{canton's zip code}
}
\value{
canton's zip code
}
\description{
Return in which canton the person is retiring
}
