% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MessageData.R
\name{UI_Messages}
\alias{UI_Messages}
\title{IB and VM list}
\usage{
UI_Messages()
}
\value{
Ib and VM list
}
\description{
list of variables within Ibfo Box and Validity Message lists
}
\examples{
\dontrun{
  # After edits, update the data e.g. by running: 
  VM <- UI_Messages()$VM
  save(VM, file = "data/VM.rda")
  # load("data/VM.rda")
}
#' @export
}
