% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{update_min}
\alias{update_min}
\title{update_min}
\usage{
update_min(input, name, session)
}
\arguments{
\item{input}{numericInput field.}

\item{session}{Current session.}
}
\value{
Zero value.
}
\description{
Automatically updates numericInput to zero if input is negative.
}
