% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{update_neg}
\alias{update_neg}
\title{update_neg}
\usage{
update_neg(inputId, session)
}
\arguments{
\item{inputId}{Field name.}

\item{session}{Current session.}
}
\value{
Zero value.
}
\description{
Automatically updates numericInput to zero if input is negative.
}
