% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withModalSpinner.R
\name{withModalSpinner}
\alias{withModalSpinner}
\title{Modal spinner information message}
\usage{
withModalSpinner(expr, info, spinner = icon("spinner", "fa-spin"),
  size = "m")
}
\arguments{
\item{expr}{The \code{expression} to be evaluated.}

\item{info}{The information message to be displayed.}

\item{spinner}{The spinning wheel icon.}

\item{size}{One of \code{"s"} for small, \code{"m"} (the default) for medium,
or \code{"l"} for large.}
}
\description{
Display a modal window with a spinning wheel and an information message
while a (time-consuming) expression is evaluated.
}
\examples{
if (interactive()) {
  library(shiny)
  ui <- fluidPage(
    titlePanel("withModalSpinner"),
    actionButton("trigger_runif", "Trigger...", icon("cog")),
    actionButton("trigger_expr", "Trigger...", icon("cog"))
  )

  slow_runif <- function(time) c(Sys.sleep(time), runif(1))
  server <- function(input, output) {
    observeEvent(input$trigger_runif, {
      withModalSpinner(
        message(slow_runif(2)),
        "Waiting for a slow runif...",
        size = "s"
      )
    })
    observeEvent(input$trigger_expr, {
      withModalSpinner(
        {
          Sys.sleep(1)
          for (i in 1:10) {
            message(i)
            Sys.sleep(0.1)
          }
        },
        "Waiting for a slow expression..."
      )

    })
  }
  # Run the application
  shiny::runApp(shiny::shinyApp(ui = ui, server = server), launch.browser = TRUE)
}
}
