% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{buildContributionP2Path}
\alias{buildContributionP2Path}
\title{Build Contribution Pillar II Path}
\usage{
buildContributionP2Path(birthday, Salary, SalaryGrowthRate, CurrentP2,
  P2purchase, TypePurchase, rate = BVGMindestzinssatz,
  givenday = today("UTC"), RetirementAge)
}
\arguments{
\item{birthday}{date of birth ("YYYY-MM-DD").}

\item{Salary}{stipend}

\item{SalaryGrowthRate}{rate at which the salary increases/decreases}

\item{CurrentP2}{TODO-Gabriel}

\item{P2purchase}{TODO-Gabriel}

\item{TypePurchase}{TODO-Gabriel}

\item{rate}{male or female}

\item{givenday}{date at which the age is computed ("YYYY-MM-DD").}

\item{RetirementAge}{age of retirement}
}
\value{
TODO-Gabriel
}
\description{
TODO-Gabriel
}
\examples{
\dontrun{
buildContributionP2Path(
  birthday = "1975-10-10",
  Salary = 90000,
  SalaryGrowthRate = 0.01,
  CurrentP2 = 10000,
  P2purchase = 2000,
  TypePurchase = "AnnualP2",
  rate = 0.025,
  givenday = as.Date("2018-07-04"),
  RetirementAge = 67)
}
}
