% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{buildContributionP3path}
\alias{buildContributionP3path}
\title{Build the contribution path for a standard pension fund, called Pillar 3 in Switzerland.
Based on 'calcAnnuityAcumPath()'}
\usage{
buildContributionP3path(birthday, P3purchase, CurrentP3, returnP3,
  givenday = today("UTC"), RetirementAge)
}
\description{
Build the contribution path for a standard pension fund, called Pillar 3 in Switzerland.
Based on 'calcAnnuityAcumPath()'
}
\examples{
\dontrun{
buildContributionP3path(
  birthday = "1980-12-01",
  P3purchase = 5000,
  CurrentP3 = 100000,
  returnP3 = 0.03,
  givenday = as.Date("2015-11-30"),
  RetirementAge = 62)
}
}
\concept{Contributionpath}
