% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{buildt}
\alias{buildt}
\title{Buildt}
\usage{
buildt(birthday, givenday = today("UTC"), RetirementAge = 65)
}
\arguments{
\item{birthday}{Date of birth ("YYYY-MM-DD").}

\item{givenday}{Date at which the age is computed ("YYYY-MM-DD").}

\item{RetirementAge}{Age of retirement}
}
\value{
Vector of dates until retirement.
}
\description{
Build annual path from today (or given day) until the retirement day.
}
\examples{
\dontrun{
buildt("1981-08-12")
}
}
