% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxBenefit.R
\name{buildTaxBenefits}
\alias{buildTaxBenefits}
\title{Builds a data frame with the tax benefits path}
\usage{
buildTaxBenefits(birthday, TypePurchase, P2purchase, P3purchase, returnP3,
  Salary, SalaryGrowthRate, postalcode, NKids, churchtax, rate_group,
  givenday = today("UTC"), RetirementAge)
}
\description{
Builds a data frame with the tax benefits path
}
\details{
All inputs are scalars. Builds a data frame as long as the years to retirement.
Calls 'getTaxAmount()' through 'calcTaxBenefitSwiss()', therefore, it assumes objects on the global enviornment.
}
\examples{
\dontrun{buildTaxBenefits(
 birthday,
 TypePurchase,
 P2purchase,
 P3purchase,
 returnP3,
 Salary,
 SalaryGrowthRate,
 postalcode,
 NKids,
 churchtax,
 rate_group,
 MaxContrTax,
 givenday = today("UTC"),
 RetirementAge = 65)
}
}
\seealso{
[calcTaxBenefitSwiss()]

[getTaxAmount()]

Other swisstax: \code{\link{calcTaxBenefitSwiss}},
  \code{\link{getTaxAmount}}, \code{\link{lookupTaxAmount}}
}
\concept{swisstax}
