% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxBenefit.R
\name{buildTaxBenefits}
\alias{buildTaxBenefits}
\title{Builds a data frame with the tax benefits path}
\usage{
buildTaxBenefits(birthday, TypePurchase, P2purchase, P3purchase, returnP3,
  Salary, SalaryGrowthRate, postalcode, NKids, churchtax, rate_group,
  givenday = today("UTC"), RetirementAge)
}
\arguments{
\item{birthday}{date of birth ("YYYY-MM-DD").}

\item{TypePurchase}{TODO-Gabriel}

\item{P2purchase}{Pillar II purchase}

\item{P3purchase}{Pillar III purchase}

\item{returnP3}{TODO-Gabriel}

\item{Salary}{stipend}

\item{SalaryGrowthRate}{rate at which the salary increases/decreases}

\item{postalcode}{zip code \code{Character}}

\item{NKids}{number of children \code{Numeric} scalar.}

\item{churchtax}{TODO-Gabriel \code{Character} Y/N}

\item{rate_group}{TODO-Gabriel. \code{Character}.}

\item{givenday}{date at which the age is computed ("YYYY-MM-DD").}

\item{RetirementAge}{age of retirement}
}
\value{
TODO-Gabriel
}
\description{
Builds a data frame with the tax benefits path
}
\details{
All inputs are scalars. Builds a data frame as long as the years to retirement.
Calls 'getTaxAmount()' through 'calcTaxBenefitSwiss()', therefore, it assumes objects on the global enviornment.
}
\examples{
\dontrun{buildTaxBenefits(
 birthday,
 TypePurchase,
 P2purchase,
 P3purchase,
 returnP3,
 Salary,
 SalaryGrowthRate,
 postalcode,
 NKids,
 churchtax,
 rate_group,
 MaxContrTax,
 givenday = today("UTC"),
 RetirementAge = 65)
}
}
\seealso{
Other swisstax: \code{\link{calcTaxBenefitSwiss}},
  \code{\link{getTaxAmount}}, \code{\link{lookupTaxAmount}}
}
\concept{swisstax}
