% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxBenefit.R
\name{buildTaxBenefits}
\alias{buildTaxBenefits}
\title{buildTaxBenefits}
\usage{
buildTaxBenefits(birthday, TypePurchase, P2purchase, P3purchase, returnP3,
  Salary, SalaryGrowthRate, postalcode, NChildren, churchtax, rate_group,
  givenday = today("UTC"), RetirementAge)
}
\arguments{
\item{birthday}{Date of birth ("YYYY-MM-DD").}

\item{TypePurchase}{AnnualP2 if the purchase on the second pillar gets repeated every year until retirement.}

\item{P2purchase}{Annual voluntary contribution to the Occupational Fund.}

\item{P3purchase}{Annual voluntary contribution to the Private Fund (3rd Pillar).}

\item{returnP3}{Annual expected return on the Private Fund.}

\item{Salary}{Stipend.}

\item{SalaryGrowthRate}{Rate at which the salary increases/decreases.}

\item{postalcode}{Zip code \code{Character}}

\item{NChildren}{Number of children. \code{Numeric} scalar.}

\item{churchtax}{Y/N \code{Character} Y/N}

\item{rate_group}{A (Single), B (Married), C (Married Double income) \code{Character}.}

\item{givenday}{Date at which the age is computed ("YYYY-MM-DD").}

\item{RetirementAge}{Age of retirement.}
}
\value{
data.frame tax benefits path.
}
\description{
All inputs are scalars. Builds a data frame as long as the years to retirement.
Calls 'getTaxAmount()' through 'calcTaxBenefitSwiss()', therefore, it assumes objects on the global enviornment.
}
\examples{
\dontrun{buildTaxBenefits(
 birthday,
 TypePurchase,
 P2purchase,
 P3purchase,
 returnP3,
 Salary,
 SalaryGrowthRate,
 postalcode,
 NChildren,
 churchtax,
 rate_group,
 MaxContrTax,
 givenday = today("UTC"),
 RetirementAge = 65)
}
}
\seealso{
swisstax
}
