% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calcAnnuityAcumPath}
\alias{calcAnnuityAcumPath}
\title{TODO-Gabriel}
\usage{
calcAnnuityAcumPath(contributions, t, rate)
}
\arguments{
\item{contributions}{TODO-Gabriel}

\item{t}{TODO-Gabriel}

\item{rate}{male or female}
}
\value{
TODO-Gabriel
}
\description{
TODO-Gabriel
}
\examples{
\dontrun{
calcAnnuityAcumPath(contributions = c(50000, 1000, 1000, 1000, 1000), 
                                t = c(0.284931, 1, 1, 1, 0), rate = 0.01)
}
}
