% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calcAnnuityAcumPath}
\alias{calcAnnuityAcumPath}
\title{calcAnnuityAcumPath}
\usage{
calcAnnuityAcumPath(contributions, t, rate)
}
\arguments{
\item{contributions}{Vector of contributions (annuities).}

\item{t}{Vector of time intervals between contributions.
\itemize{
\item Irregular time intervals are allowed.
\item For frequency bellow annual, enter t as proportion of a year.
}}

\item{rate}{Interests rate on annual basis. Constant interest rates are assumed.}
}
\value{
Vector of accumulated benefits given a set of contributions.
}
\description{
Calculate future value of an annuity with periodic contributions.
\itemize{
\item Based on continuous compounding interest in annual basis.
\item Payments occur at the beginning of each period.
}
}
\examples{
\dontrun{
calcAnnuityAcumPath(contributions = c(50000, 1000, 1000, 1000, 1000),
                                t = c(0.284931, 1, 1, 1, 0), rate = 0.01)
}
}
