% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calcBVGpurchase}
\alias{calcBVGpurchase}
\title{Calculate Purchase Pilar II}
\usage{
calcBVGpurchase(TypePurchase, P2purchase, ncp)
}
\arguments{
\item{TypePurchase}{AnnualP2 if the purchase on the pillar II gets repeated every year until retirement.}

\item{P2purchase}{Annual voluntary contribution to the Occupational Pension Fund.}

\item{ncp}{Length contribution path to retirement}
}
\value{
BVG purchase
}
\description{
Calculate the path of purchases to the Pilar II (Occupational pension fund, BVG)
}
\examples{
\dontrun{
calcBVGpurchase(TypePurchase = "AnnualP2", P2purchase = 2000, ncp = 25) \%>\% print
}
}
