% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{calcExpectedSalaryPath}
\alias{calcExpectedSalaryPath}
\title{Calculate Expected Salary Path}
\usage{
calcExpectedSalaryPath(Salary, SalaryGrowthRate, ncp)
}
\arguments{
\item{Salary}{stipend}

\item{SalaryGrowthRate}{rate at which the salary increases/decreases}

\item{ncp}{length contribution path to retirement}
}
\value{
expected salary path
}
\description{
calculate whether the salary will increase/decrease and by how much.
}
\examples{
\dontrun{
calcExpectedSalaryPath(90000, 0.02, 20)
}
}
