% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxBenefit.R
\name{calcTaxBenefitSwiss}
\alias{calcTaxBenefitSwiss}
\title{Calculates the tax benefit as a difference of the taxes paid with and without retirement contributions.}
\usage{
calcTaxBenefitSwiss(ExpectedSalaryPath, TaxableIncome, rate_group, Age,
  NKids, postalcode, churchtax)
}
\arguments{
\item{ExpectedSalaryPath}{vector length equals year to retirement}

\item{TaxableIncome}{vector length equals year to retirement}

\item{Age}{vector length equals year to retirement}
}
\description{
Calls 'getTaxAmount()', therefore, it assumes objects in the global environment.
}
\examples{
\dontrun{
  calcTaxBenefitSwiss(ExpectedSalaryPath = seq(90000, 100000, 1000),
                    TaxableIncome = seq(88000, 98000, 1000),
                    rate_group = "A",
                    Age = seq(55, 65),
                    NKids = 0,
                    postalcode = 8400,
                    churchtax = "Y")
}
}
\seealso{
[getTaxAmount()]

Other swisstax: \code{\link{buildTaxBenefits}},
  \code{\link{getTaxAmount}}, \code{\link{lookupTaxAmount}}
}
\concept{swisstax}
