% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxBenefit.R
\name{calcTaxBenefitSwiss}
\alias{calcTaxBenefitSwiss}
\title{calcTaxBenefitSwiss}
\usage{
calcTaxBenefitSwiss(ExpectedSalaryPath, TaxableIncome, rate_group, Age,
  NChildren, postalcode, churchtax)
}
\arguments{
\item{ExpectedSalaryPath}{Vector of annual salaries until retirement.}

\item{TaxableIncome}{Vector of annual taxable income until retirement.}

\item{rate_group}{A (Single), B (Married), C (Married Double income) \code{Character}.}

\item{Age}{Age of the person. \code{Numeric}}

\item{NChildren}{Number of children. \code{Numeric} scalar.}

\item{postalcode}{Zip code \code{Character}}

\item{churchtax}{Y/N \code{Character} Y/N}
}
\value{
Single tax benefits (tax relief) of one contribution.
}
\description{
Calculates the tax benefits as a difference of the taxes paid with and without retirement contributions.
Calls 'getTaxAmount()', therefore, it assumes objects in the global environment.
}
\examples{
\dontrun{
  calcTaxBenefitSwiss(ExpectedSalaryPath = seq(90000, 100000, 1000),
                    TaxableIncome = seq(88000, 98000, 1000),
                    rate_group = "A",
                    Age = seq(55, 65),
                    NChildren = 0,
                    postalcode = 8400,
                    churchtax = "Y")
}
}
\seealso{
\code{\link[=getTaxAmount]{getTaxAmount()}}

swisstax
}
