% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{changeToPercentage}
\alias{changeToPercentage}
\title{changeToPercentage}
\usage{
changeToPercentage(df)
}
\arguments{
\item{df}{Given data frame.}
}
\value{
Percentage value.
}
\description{
From decimal to percentage value.
}
