% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{getRetirementCalendar}
\alias{getRetirementCalendar}
\title{Obtain Retirement Calendar}
\usage{
getRetirementCalendar(birthday, givenday = today("UTC"),
  RetirementAge = 65)
}
\arguments{
\item{birthday}{Date of birth ("YYYY-MM-DD").}

\item{givenday}{Date at which the age is computed ("YYYY-MM-DD").}

\item{RetirementAge}{Age of retirement}
}
\value{
retirement calendar
}
\description{
Calculate the annual retirement path.
SmarP assumes that the user will get retired the day that turns 65 or its desired retirement age.
For the current year, if birthday is later than the calculation day, there will be 2 dates.
}
\examples{
\dontrun{
getRetirementCalendar("1981-08-12")
getRetirementCalendar("1981-08-12", as.Date("2018-12-02"), RetirementAge = 62)
}
}
