% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxBenefit.R
\name{getTaxAmount}
\alias{getTaxAmount}
\title{getTaxAmount}
\usage{
getTaxAmount(Income, rate_group, Age, NChildren, postalcode, churchtax)
}
\arguments{
\item{Income}{Annual salary. \code{Numeric} scalar.}

\item{rate_group}{A (Single), B (Married), C (Married Double income) \code{Character}.}

\item{Age}{Age of the person. \code{Numeric}}

\item{NChildren}{Number of children. \code{Numeric} scalar.}

\item{postalcode}{Zip code \code{Character}}

\item{churchtax}{Y/N \code{Character} Y/N}
}
\value{
Tax Amount
}
\description{
This function uses 2 main sources for tax data.
At Kanton and Gemeinde level, the source is taxburden.list.
At federal level, we use the official taxrate table (BundessteueTabelle) and we try to aproximate the taxable income.
}
\details{
This function assumes the following objects on the global enviornment
\itemize{
\item PLZGemeinden (includes Kirchensteuer)
\item taxburden.list
\item BundessteueTabelle
\item BVGcontributionratesPath, BVGcontributionrates
\item MaxBVG, MinBVG
\item KinderabzugKG
\item NBU, maxNBU
\item AHL
\item ALV, maxALV
\item VersicherungsL, VersicherungsV, VersicherungsK
\item BerufsauslagenTarif, BerufsauslagenMax, BerufsauslagenMin
}
}
\examples{
\dontrun{
getTaxAmount(Income = 200000, rate_group = "C", Age = 32,
             NChildren = 5, postalcode = 8400, churchtax = "Y")
}
}
\seealso{
swisstax
}
