% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{isnotAvailable}
\alias{isnotAvailable}
\title{isnotAvailable}
\usage{
isnotAvailable(inputValue)
}
\arguments{
\item{inputValue}{Input value.}
}
\value{
TRUE or FALSE.
}
\description{
If input value is not available, then return logical TRUE else FALSE.
}
