% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{isnotAvailableReturnZero}
\alias{isnotAvailableReturnZero}
\title{Is Not Available Return Zero}
\usage{
isnotAvailableReturnZero(inputValue, fallback = 0)
}
\arguments{
\item{inputValue}{input value}

\item{fallback}{zero}
}
\value{
zero or input value
}
\description{
if input value is not available, then return zero else input value
}
