% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_application.R
\name{launch_application}
\alias{launch_application}
\title{SmaRP App Launcher}
\usage{
launch_application(...)
}
\arguments{
\item{...}{Additional arguments passed on to \code{\link[=runApp]{runApp()}}. Note that
argument \code{launch.browser} is always passed as \code{TRUE}.}
}
\value{
Side-effecting function. Launches the SmaRP app.
}
\description{
Launch the SmaRP Shiny wep app.
}
\examples{
if (interactive()) {
  SmaRP::launch_application()
}
}
