% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_application.R
\name{launch_application}
\alias{launch_application}
\title{SmaRP App Launcher}
\usage{
launch_application(...)
}
\arguments{
\item{...}{additional arguments to be passed on to runApp}
}
\value{
Side-effecting function. Launches SmaRP application.
}
\description{
function to launch the SmaRP shiny app
}
\examples{
\dontrun{
  library(SmaRP); launch_application()
}
}
