% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxBenefit.R
\name{lookupTaxAmount}
\alias{lookupTaxAmount}
\title{Returns the tax amount to be paid given one income.}
\usage{
lookupTaxAmount(Income, Tabelle, CivilStatus)
}
\arguments{
\item{Income}{annual stipend}

\item{Tabelle}{Income - Tax rate table at Federal level.}

\item{CivilStatus}{marital status}
}
\value{
tax amount to be paid
}
\description{
Search the tax amount to be paig given one income on the tax tables.
}
\examples{
\dontrun{
lookupTaxAmount(Income = 100000, Tabelle = BundessteueTabelle, CivilStatus = "A")
}
}
\seealso{
Other swisstax: \code{\link{buildTaxBenefits}},
  \code{\link{calcTaxBenefitSwiss}},
  \code{\link{getTaxAmount}}
}
\concept{swisstax}
