% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxBenefit.R
\name{lookupTaxAmount}
\alias{lookupTaxAmount}
\title{Returns the tax amount to be paid given one income.}
\usage{
lookupTaxAmount(Income, Tabelle, CivilStatus)
}
\description{
Returns the tax amount to be paid given one income.
}
\examples{
\dontrun{
lookupTaxAmount(Income = 100000, Tabelle = BundessteueTabelle, CivilStatus = "A")
}
}
\seealso{
Other swisstax: \code{\link{buildTaxBenefits}},
  \code{\link{calcTaxBenefitSwiss}},
  \code{\link{getTaxAmount}}
}
\concept{swisstax}
