% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{makeTable}
\alias{makeTable}
\title{Make Table}
\usage{
makeTable(Road2Retirement, moncols = c("DirectP2", "ReturnP2", "TotalP2",
  "DirectP3", "ReturnP3", "TotalP3", "DirectTax", "ReturnTax", "TotalTax",
  "Total"))
}
\arguments{
\item{Road2Retirement}{TODO-Gabriel}

\item{moncols}{TODO-Gabriel}
}
\value{
TODO-Gabriel
}
\description{
TODO-Gabriel
}
\examples{
\dontrun{
makeTable(Road2Retirement)
}
}
