% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{need_not_zero}
\alias{need_not_zero}
\title{Need Not Zero}
\usage{
need_not_zero(input, inputname)
}
\arguments{
\item{input}{zero, nothing or null}

\item{inputname}{name of input}
}
\value{
TODO-Gabriel
}
\description{
TODO-Gabriel
}
