% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{returnPLZKanton}
\alias{returnPLZKanton}
\title{returnPLZKanton}
\usage{
returnPLZKanton(plz)
}
\arguments{
\item{plz}{Canton's postal code.}
}
\value{
Canton.
}
\description{
Return in which canton the person is retiring.
}
